+++ NEWS.md
@@ -27,14 +27,14 @@
 * Add min/max columns to THETA's #24
 * Method 'addSuffix' does not preserve the variance-covariance matrix #81
 * Method 'add' does not merge variance-covariance matrices #82
-* Method 'addRSE' to quickly add relative standard error on parameter #83
+* Method 'add_rse' to quickly add relative standard error on parameter #83
 * Implement generic method 'move' #84
 * Generate parameter uncertainty in campsismod rather than in campsis #85
 * Error when printing the model in the console #88
 * Allow manual import of sampled parameters for model replication #89
 * Get rid of plyr package #90
 * Quality: increase code coverage #91
-* Method 'addRSE' can't be used to replace an existing value #92
+* Method 'add_rse' can't be used to replace an existing value #92
 * Issue when the model is replicated only once #93
 * Implement show method on replication setting object #95
 * Check OMEGA and SIGMA matrix for positive definiteness #96